Migrate template transformations to `filters` module
